app/grade_service/requirements.txt
flask
Flask-SQLAlchemy
flask_sqlalchemy
requests